REF: Move Output Manager to its own dedicated file
